typescript issue fixes
