+++ frontend/client/src/components/navbar/data/index.tsx
@@ -1,10 +1,12 @@
 import {
+  ChatIcon,
   GiftIcon,
   MakeupKitIcon,
+  PlayIcon,
   TestIcon,
   ValueSetsIcon,
 } from "../../../icons";
-import { LevelOneCategoryType } from "../../../types";
+import { LevelOneCategoryType } from "../types";
 import About from "../components/about/About";
 import Collections from "../components/Collections";
 import Eyes from "../components/Eyes";
@@ -47,6 +49,105 @@ import {
 } from "../components/icons";
 import Lips from "../components/lips/Lips";
 import Skin from "../components/skin/Skin";
+import { SocialCommunityItem } from "../types";
+import { TextItem } from "../../../types";
+
+export const socialCommunity: SocialCommunityItem[] = [
+  {
+    id: 1,
+    icon: PlayIcon,
+    label: "Founder's Story: Watch Now",
+    link: "https://www.youtube.com/watch?v=92kcChL74ZE",
+  },
+  {
+    id: 2,
+    icon: ChatIcon,
+    label: "Chat with our team",
+    link: "/contact-us",
+  },
+];
+
+export const lipstickFeedbacks: TextItem[][] = [
+  [
+    { text: "Great variety of lipsticks! Loved the" },
+    { text: "shades", isHighlighted: true },
+    { text: "and" },
+    { text: "textures.", isHighlighted: true },
+    { text: "Long-lasting", isHighlighted: true },
+    {
+      text: "and comfortable. Perfect for every occasion. Highly recommended!",
+    },
+  ],
+  [
+    { text: "Amazing" },
+    { text: "pigmentation", isHighlighted: true },
+    { text: "and" },
+    { text: "smooth application.", isHighlighted: true },
+    { text: "Stays on for hours" },
+    { text: "without drying out lips. Definitely a" },
+    { text: "must-have!", isHighlighted: true },
+    { text: "Long-lasting and" },
+    { text: "comfortable.", isHighlighted: true },
+  ],
+  [
+    { text: "Beautiful" },
+    { text: "shades", isHighlighted: true },
+    { text: "and great" },
+    { text: "texture!", isHighlighted: true },
+    { text: "Perfect" },
+    { text: "matte", isHighlighted: true },
+    { text: "finish without feeling heavy. Love the" },
+    { text: "Long-lasting", isHighlighted: true },
+    { text: "effect! Amazing quality and" },
+    { text: "pigmentation.", isHighlighted: true },
+  ],
+
+  [
+    { text: "Excellent" },
+    { text: "color payoff!", isHighlighted: true },
+    { text: "Super" },
+    { text: "comfortable", isHighlighted: true },
+    { text: "to wear all day. My" },
+    { text: "go-to lipstick", isHighlighted: true },
+    {
+      text: "for every event. Worth every penny! Absolutely stunning luxurious.",
+    },
+  ],
+  [
+    { text: "Lovely" },
+    { text: "shades", isHighlighted: true },
+    { text: "with a" },
+    { text: "creamy texture.", isHighlighted: true },
+    { text: "No" },
+    { text: "smudging", isHighlighted: true },
+    {
+      text: "and lasts all day. Received so many compliments. Absolutely love them! Truly fantastic.",
+    },
+  ],
+
+  [
+    { text: "The" },
+    { text: "colors", isHighlighted: true },
+    { text: "are vibrant and bold. Great for all" },
+    { text: "skin tones.", isHighlighted: true },
+    { text: "Stays" },
+    { text: "intact", isHighlighted: true },
+    { text: "and flawless" },
+    { text: "even after meals. Highly recommended! Truly amazing." },
+  ],
+  [
+    { text: "Impressive" },
+    { text: "quality", isHighlighted: true },
+    { text: "and" },
+    { text: "shade range.", isHighlighted: true },
+    { text: "Glides" },
+    { text: "so" },
+    { text: "smoothly", isHighlighted: true },
+    {
+      text: "and feels so lightweight. Absolutely perfect for both daily wear and special occasions!",
+    },
+  ],
+];
 
 export const for_you: LevelOneCategoryType = {
   id: 1,
